feat(radio/tx): updated av_state.cpp
Following changes were made :
- Updated data dump to better reflect AV state
- Updated av_state.cpp to reuse the new data dump values

Misc changes :
- Remove some trivial warnings flooding compilation console (mostly in manual tests)

+++ Drivers/Buzzer/Tests/Hardware/buzzer_manual_test.cpp
@@ -10,7 +10,7 @@ extern "C" {
 void manual_test_buzzer_set_buzzer (bool status) {
     HAL_GPIO_WritePin(BUZZER_GPIO_Port, BUZZER_Pin, status ? GPIO_PIN_SET : GPIO_PIN_RESET);
 
-    printf("[BUZZER] Set at time %d: %d\r\n", HAL_GetTick(), status);
+    printf("[BUZZER] Set at time %lu: %d\r\n", HAL_GetTick(), (int) status);
 }
 void manual_test_buzzer () {
     bool started = false;
